requirements.txt
scrapy>=2.11
beautifulsoup4>=4.12
lxml>=5.1
pyyaml>=6.0
matplotlib>=3.8
numpy>=1.26
Pillow>=10.2
